backend/requirements.txt
pyadi-iio>=0.0.15
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
# gnuradio  # System dependency
websockets>=12.0
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
uvicorn>=0.25.0
fastapi>=0.109.0
